Fixed two bugs in the control-flow checker dealing with statement-expressions.
First, the code was ignoring flows along non-local-predecessors, hence
it looked like code after a switch in (the "top" of) a
statement-expression was unreachable.  (Follow that? :-) )

Second, because of statement-expressions and local fixpoint
computations on expressions, it is not the case that each statement is
analyzed exactly once per global iteration.  Hence the visited field
should not be incremented; it should be updated to the
iteration-number.

No more warnings for tools/cyclex/compact.cyc