* bzero was not quite right for some reason.  Fixed.
* Added --nochecks, --nonullchecks, --noboundschecks; modified cyc_include to ifdef these checks out if the corresponding command line options are enabled.  As a result every genfile has changed.


* Changed memcpy, memmove, bcopy to use the #define hack e.g.

extern `a::A?`r _memcpy(`a?`r d, const `a? s, size_t, sizeof_t<`a>);

#define _rounddiv(q,d) (((q)/(d)) + ((((q)%(d)) == 0) ? 0 : 1))

#define memcpy(d,s,n) _memcpy((d),(s),_rounddiv(n,sizeof(*(s))),sizeof(*(s)))

The *s "dereference" is safe because sizeof does not evaluate its expression arguments.

Perhaps _memcpy etc. should not be exposed in stdlib.h.
Perhaps it should be a run time error to attempt a partial copy rather than a silent successful full copy of the structure you attempted to partially copy.