Got the order of arguments to calloc wrong.  Because this had effects on cyc_include.h, I ended up changing everything (sorry).
The tests/shootout programs were minimized for diffs by
using malloc/calloc where appropriate and now I am quite
happy with the performance we're seeing.  For example,
reversefile is only 50% slower (it was a *lot* slower before.)
The major change here is that we're using realloc for
the buffer, just like the C code.

I do note that compiling these with nocheck does not
give major improvements in performance for most benchmarks.