* Fixed the parser to use Tcutil::warn instead of its own function.
* We now only warn when you don't specify a type in a declarator
and substitute "int" for it.  This means you can leave return types
off and they'll default to int (but you'll get a warning).

* We now only warn (instead of generating an error) when you fail
to return a char, short, int, float, or double.

* We now enforce and use the "noreturn" attribute on functions.
This means that you can call a function such as exit() or abort()
and the control-flow analysis treats it the same as throwing an
exception.  It also means that you can write your own "noreturn"
functions, but the analysis checks to make sure you don't explicitly
or (implicitly) return.

* Simplified the assert macro and __assert_fail definition by
using the above facilities (i.e., removed a gratuitous `throw'.)

* Updated todo list a bit more.