Some more tweaks for smaller output (another 390K):
* Just put int tags for tunions directly in the code rather than
generating static const int variables.  This is perhaps controversial
because it makes it harder to debug the C code.  I argue that the C
code for tunions was already unreadable by mortals.  This is a double
win (and accounts for most of the 390K): we don't generate globals
which often aren't used, and we have smaller code for manipulating tags.


* We weren't tree-shaking under extern "C" { ... }

* slight change to how we cast in pattern translation

* _LL and _tmp now use different counters and _LL also uses %X

Also exploited the flow-analysis's knowledge of noreturn in toc.cyc.