Lots of small changes.  Only "external" change is that struct-building expressions can now have explicit witness types, such as Fn{<int> take_int, 0}.  You have to use curly braces if you have explicit witnesses.
Regionified the code for checking pattern-exhaustiveness.  For
compiling tcutil.cyc, that cuts the number of calls to GC_malloc by
about 10 percent.

Changed the translation of patterns to not emit labels for the
"success path".  gcc -O recovers this, but I think it was actually
cleaner to make this change -- xlate_pat never has to generate fresh
labels.  In a related change, toc.cyc no longer uses the exhaust field
on Let_d, but I haven't removed it yet.