* Made the assignment-in-test warning in marshal go away
* Added "register" storage class.  Previously we were ignoring this
(turning it into Public).  Now we pass it through to the C compiler.

* Added a flag "-noregister" that restores the old behavior of ignoring
register storage class, for performance comparisons.

* Made it a warning instead of an error for main to return void, lots
of benchmarks do this, so it saves us a frivolous diff.

Libraries printf and getopt, as well as Cyclone bison, use the register
keyword.  So do many benchmarks.  gcc -O0 does maake use of register
scope declarations.  But, gcc -O3 either ignores, or its analysis
supercedes, register declarations.  So I haven't seen any actual
performance improvements as a result of this, but it does no harm either.

C complains that we're taking the address of a register variable whenever
we do inplace pointer arithmetic on a fat pointer declared register,
because the in-place plus works indirectly through the address of
the tag_arr struct.  This is really silly since structs can't fit into
registers anyway, so perhaps I should fix it so that register declarations
of fat pointers get filtered out.