* The massive changes are due to my last checkin which changed cyc_include.h.
* Fixed bug: locally declared zero-terminated arrays must be
initialized now.  More generally, the flow analysis does not
consider a zero-terminated array to be "bits-only".  We may
relax this and put in an implicit 0 at the end of the array
for convenience in the future.

* Fixed bug in libc.cys for sys/time.h on Cygwin.  Or more
properly, hacked around the fact that fdset is a macro on
Cygwin that's set to _types_fd_set.  The problem is that
we were missing the "typedef struct _types_fd_set _types_fd_set"
needed.  We already "fixed" this with an epilogue for
socket.h, but it's also needed for time.h so I did what I
could.  Hope this doesn't break linux.

* Fixed tests/shootout so that they now compile again.