First, the trivial stuff: * changed the assert in string.cyc, strcmp so that Greg's array-bound stuff can exploit it. * made some static global variables static local variables since we now do the right thing for these
Second, some stuff that changed as a result of the subscript-checking
for abstract-length arrays:
* changed translation of tag_t<t> to unsigned int.  This is what you
want so that you get more unsigned comparisons.
* changed how the control-flow analysis evaluates conditional
expressions.  The new way is a bit slower and a bit less conservative.

Now, for subscript-checking: given a pointer of type "t @{`i}" (or t
*{`i}) you can actually index into it.  (I don't yet support * or ->
though.)  There is never a run-time check -- the flow analysis must be
able to prove that the index is in-bounds.  The simplest thing to do
is compare an "unescaped" integer against a tag indicating the size of
the array.  Here's a simple example:

int sum(tag_t<`a> sz, int@ @{`a} arr) {
  int ans = 0;
  for(int i=0; i < sz; ++i)
    ans += *arr[i];
  return ans;
}

However, the flow analysis is much less ad hoc than this suggests.
For those of you who have dared to try to understand the flow
analysis, there's a new "abstract rvalue" that is a list of "tag
comparisons" such as "equal to `a" or "less than `1000".  Then a test
of the form "i < sz" propagates these tag comparisons from sz to i.
All 6 comparisons are handled -- if i already had tag comparisons,
they would propagate to sz.

Now I'll show off:

int my_bufcmp(tag_t<`i> sz1, tag_t<`j> sz2, int @{`i} s1, int @{`j} s2) {
  int min = (sz1 < sz2) ? (unsigned)sz1 : sz2;
  unsigned i = 0;
  while(i < min) {
    int c1 = s1[i];
    int c2 = s2[i];
    if(c1 != c2)
      return c1-c2;
    ++i;
  }
  return 0;
}

The cast -- (unsigned)sz1 -- is the fault of tcConditional and isn't
really the issue.  The code is accepted.  On function entry we know
sz1: = `i
sz2: = `j.
So after (sz1 < sz2) we propagate
to the true side:
   sz1: =`i, <`j
   sz2: =`j
and to the false side:
   sz1: =`i
   sz2: =`j, <=`i
so when joining this information, we'll propagate
   sz1: =`i
   sz2: =`j
   min: <=`i, <=`j  **this is the cool part**
The rest follows easily because i < min is an unsigned comparison.
Note the function will fail to check if i has type int or you have an
off-by-one error (e.g. i <= min).

Finally, there are enough silent casts that we can write:
int main() {
  int @x[100] = {for i < 100: (new ((int) i))};
  sum(100,x);
  return 0;
}

The biggest shortcoming is the lack of arithmetic.  So arr[i-1] is
never going to work.  I also don't support tests like (++i < min)
or (i++ < min), but I think I'll be able too, especially the latter
one.

I've kept all this stuff quite separate from everything else, in
particular the array-bounds elimination for other pointer types.  So
the fact that everything still bootstraps is pretty trivial -- nothing
in the libraries or compilers will trigger any of my new code.