One major and many minor improvements that I wanted to sneak in before the release.
* got rid of warnings that you are dereferencing a possibly null
pointer when you've already checked for null.  For instance:

  FILE *fp = fopen(f,m);
  if (fp == NULL) fp = stdin;
  fgets(buffer,N,fp);

no longer gives a warning that fp might be null upon the call to
fgets.  I did this by modifying the type-checker so that it records
what kind of casts it inserts and remains silent when we have to
convert a t* to a t@.  The flow analysis then treats a cast from t* to
t@ as a potential null check.  That means that (a) it tries to
determine whether or not the check is necessary and (b) records for
the continuation that the value is definitely not null.  Then
subsequent null checks on the same value are no longer needed.
Finally, when doing code generation, we issue a warning if the
compiler had to insert a cast from t* to t@ that resulted in a null
check.  This change has the added advantage that it eliminates some
unnecessary null-checks that we left in when the user explicitly cast
from t* to t@.

* made control-flow analysis a little more accurate.  It now detects
dead code for conditionals that test for e1 == e2 or e1 != e2 when
we can figure out whether or not e1 and e2 are zero.  This helps
avoid flowing through certain impossible cases in conditionals
and makes the flow analysis a little more accurate.

* made control-flow analysis understand @ types.  In particular,
when doing the abstract evaluation of an expression of type t@,
we know that the result is at least not zero.  This is needed to
get the "fp = stdin" to propagate the fact that fp can't be null
after the assignment.

* fixed scanf so that you can pass in either a ZEROTERM or NOZEROTERM
char pointer for %s, and fixed the type-checkers conversions so that
it defaults to NOZEROTERM.  This allows more buffers to be passed in
to scanf.

* by default, we no longer print effects for function types.  This can
be overridden with a command-line flag.  Improves error messages for
functions considerably.

* expr_err takes in the topt and if non-NULL, returns that expected
type instead of a unification variable which gets rid of some cascading
error messages.  This works particularly well when porting C code
because there's not much inference going on (i.e., we usually
have a topt.)

* better error reporting when a member or argument has a qualifier
mismatch.  The member/argument is identified as the culprit instead
of the enclosing struct/function type.

* some subtyping on functions added -- generally, the subtyping
machinery needs to be beefed up to take advantage of this, but
I was hacking in that general area so I thought I'd throw it in.

* we print a warning now when you cast the result of a malloc/calloc
that's returning a t? to a t*.  This helps when porting legacy code
because it will inevitably have such a cast, and then all of the
size information gets lost.