* Factored boundscmp into a pure comparison (boundscmp) that only typecmp uses, and a unify/compare function (unify_it_bounds) that everything else uses.  This gets rid of a lot of spurious "different pointer bounds" errors, e.g. when you want to convert a char@{`45} to/from a char@{45}.
* Made it so that casting from an abstract upper bound to another
abstract upper bound is allowed in toc rather than raising a toc::impos.
This allows, for example, casting a char*{`i} to a char@{`i} (which
inserts a null check) and back (which doesn't).  Typechecking/unification
is (or at least should be) checking that the abstract bounds match up
(i.e. not allowing casting char@{`i} to char@{`j} willy nilly).