A bunch of changes to admit building on FreeBSD (at least the configuration for some machines in the undergrad lab); points 1 and 3 have action items for other folks, if they could help me out.
1) Changed realconfigure so that it no longer uses [[ ]]; use sed
instead.  (Should verify that we still get the desired effect under
Cygwin---can someone do this?)

2) Changed libc.cys slightly; needed to add off_t to the preconditions
for stdio.h, and had to add FreeBSD to the list of machines needing
the fd_set hack.

3) Changed the parser so that it accepts attributes of the form

__attribute__( ID ( ID ) )

This was to properly parse __attribute__(__mode__(__DI__)) in some
typedefs.  Buildlib already has a hack for this (discovered it
after I made the change), but I think it's buggy.  The above mode
indicates a double-sized word as can be seen from the FreeBSD def:

typedef int __attribute__((__mode__(__DI__)))            __int64_t;

We currently ignore attributes in typedefs, but this needs to be
fixed in the compiler/parser.  Judging from the current hack in
buildlib, I wouldn't be surprised if the current libraries are just
waiting to fail.  We could treat this attribute specially and
just convert the declaration into a long long (perhaps in buildlib).

4) Changed the Makefile so that there is a new target new_arch.  After
a new architecture has been created, and its genfiles were successfully
built, if you do

make UPDATEARCH=foo new_arch

where foo is your new architecture, it will create the foo directory
in bin/genfiles and copy all of the relevant files there.  It then
creates a patch from that directory.