Fixed bug in control-flow analysis on assigning to a union member. (The test in the case Greg flagged was looking for a union type, when it should look for a pointer to a union type.)
Actually, I still do the conservative thing, which is that assignment
to a union type might make the value being assigned escape (go to an
unknown location).

This is actually a less-common case of a general shortcoming in the
analysis.  There are currently two kinds of abstract lvalues:
 (1) a particular place (i.e., I know I'm assigning to x)
 (2) Unknown (don't know where I'm assigning)
In the latter case, we consider anything pointed-to to escape.  I.e.,
if x points to y and I assign x to Unknown, then y escapes, which
means we'll get real conservative about the value of y.

But that's unnecessary when, in the course of the assignment, x gets
cast to an int type.  A third lvalue UnknownInt could distinguish this
case.  But there are other places, such as when a function-argument is
implicitly cast to int, that this still wouldn't get.

This should all get fixed at the same time, and right now probably
isn't it. ;-)