Two small changes meant to make more Linux header files go through with fewer changes:
Enum declarations can now have a trailing comma after the last
element.

You can write:

  extern int foo[];

and not get an error.  The array size, however, is set to 0 and you
get a warning that this will be an (essentially) useless array.  If
you write:

  extern int foo[]ZEROTERM;

then you get an array of size 1 (since the 0 must be there) and you
can use the array (though it's expensive.)