Partially fixed the parser so that you can use typedef names as struct field names or formal parameter names.  For instance, one can write:
typedef struct foo foo;

struct foo {
  foo *foo;
};

void f(foo foo);

The fix was, as the bison info pages suggest, to duplicate a
lot of productions.  Specifically, the specifier_qualifier_list
production has been modified so that it allows at most one
typedef name in it, and the declarator and direct_declarator
productions were duplicated (to declarator_withtypedef and
direct_declarator_withtypedef) to allow a typedef_name in
the declarator.  The combination of specifier_qualifier_list and
declarator_withtypedef gets used in struct/union definitions
as well as parameter_lists.

This should allow us to parse more C header files.  The problem
is "partially" fixed for two reasons:

First, I ended up introducing a bunch more shift-reduce conflicts
as well as some reduce-reduce conflicts.  Someday, it would be nice
if someone who is better at devising grammars took a look at this
and straightened it out.

Second, we don't follow C's rules yet because once you declare
a parameter using a typedef name as in:

  void f(foo foo) {
    ...
  }

then within the body of the function, "foo" should be parsed as
an identifier and not a typedef name.  That is, the identifier
shadows the typedef name within the declared scope.  In general,
we need to modify the parser so that it knows about entering and
exiting scopes and allow for local typedef's within those scopes
as well as this weirdness.  Fortunately, this shouldn't be a problem
for parsing most header files, and it probably won't show up in
too much code.