Added a preliminary porting tool to the system (see src/port.cyc which is now part of the compiler.)
Here's how you use it.  Suppose you have a C file foo.c that you
want to port to Cyclone.  Copy the file to foo.cyc.  Then go in
and add top-level declarations:

  __cyclone_port_on__;

    ...

  __cyclone_port_off__;

around the code that you want ported.  In particular, make sure
that you don't put the header files (e.g., #include <stdio.h>)
inside the "port-on" declarations.

Now execute:

  cyclone -c -port foo.cyc > changes.txt

This will spit out to changes.txt a list of changes that the
porting tool thinks will help.  Right now, all it tries to do
is figure out whether to convert * pointers to ? pointers and
whether or not to insert "const" in various locations.  It is
often wrong but it's a start.

In addition, there is a new tool in tools/rewrite which can
take the output of the porting tools (i.e., changes.txt above)
and it will apply the changes to the file to produce
foo_new.cyc.  The rewriter is smart enough to understand
the # line directives, so this will actually produce
new files for any C header files that you #include (and
enclose in the port-on directives.)

There's still a lot to do here, but it's a start...