Just to piss Mike off :-)
Improved the porting tool.  Had to modify the abstract syntax
to put in more locaton information.  The porting tool now knows
how to do region inference and how to put in ZEROTERM in addition
to dealing with const and * vs. ?.  I've managed to use it
to port the tests/benchmark programs with only one of the
files demanding some work after the fact.

The region inference worked out really well (surprisingly).  In
particular, it figures out whether two pointer inputs to a
function must be in the same region, and whether a pointer
result is in the same region as a pointer argument and inserts
appropriate region annotations after figuring this out.