* Allow casting tunion `r1 T to tunion `r2 T when `r1 outlives `r2.
* Put the default region page size down to something very small
(20 bytes) because very few of the regions need to grow much.
(I suspect we have a lot of empty regions and the right thing to
do is to lazily allocate the first page.)

* Region-ified the type-checking environments.  This is an extremely
sophisticated use of the region mechanisms and is thus a very good
work-out for us.  In essence, the global envrionment(s) are allocated
in one region that lives throughout type-checking.  The shared portion
of a function's environment lives for throughout the type-checking
of the function, the control environment region is allocated on an as-needed
basis while type-checking the function (it changes less frequently
than the other stuff), and the rest of the function environment is
allocated in a region local to the use.  So, for instance, when you
declare a new variable, we end up allocating a new function environment
region frgn, and placing the extended function environment f in frgn.
Note that the sub-typing and existentials are used to share portions
of the old environment (e.g., the global stuff, the shared stuff,
and in this case, the ctrl-stuff, as well as the previous variable bindings.)

It's not too surprising that this is doable, since environments should
behave in a LIFO fashion.  However, it was pretty tricky to get it
organized right.  The compiler doesn't seem any slower or faster, nor
does it seem to consume much more/less memory than previously.

The key thing that I missed was not having deep subtyping for ADT's
(e.g., dict_t<t,`r1> <= dict_t<t,`r2> when `r1 outlives `r2.)  For
instance, the local variable dictionary couldn't be shared the way
I wanted it because we didn't have this kind of sub-typing.  So,
for now, I'm using a binding data structure that is essentially
a linked list, where the tl's are declared const, so as to get the
right sub-typing and thus sharing for local variable environments.