* Added "flat" tunions:
  __flat__ tunion Foo {
    A; B; C(int); D(double,double); E(int,double,char*);
  };

These are similar to normal tunions, but the representation
is something like this:

  struct Foo_A_struct {int tag;};
  struct Foo_B_struct {int tag;};
  struct Foo_C_struct {int tag; int f1;};
  struct Foo_D_struct {int tag; double f1; double f2;};
  struct Foo_E_struct {int tag; int f1; double f2; char *f3;};
  union Foo {
    struct Foo_A_struct A;
    struct Foo_B_struct B;
    struct Foo_C_struct C;
    struct Foo_D_struct D;
    struct Foo_E_struct E;
  };

Thus, an element of type tunion Foo will take up as much space
as the largest constructor (plus any padding needed to keep things
aligned.)  Anyway, if you want to use a flat tunion, then you
don't have to use "new" or "&" for the value-carrying constructors
(unlike the normal tunions.)  I plan to convert the parser so
that it uses flat tunions so as to avoid heap-allocating the
semantic values.

* fixed bug in typedef checking:  if one of the type parameters never
showed up in the type definition, then you'd get an obscure error later
on.  That's been fixed, and we now issue a warning if you fail to use
a parameter.

* we now allow the user to omit the existentially-bound type variables
in patterns -- the type-checker sticks them in.  Also, fixed problem
where too many variables in the pattern would cause the compiler to
die.

* to really take advantage of region sub-typing more things need to
be const.  We have problems initializing const things because C
complains, and you can't cast the lhs to be not-const.  So, for now,
I've changed toc.cyc to drop the const qualifier on struct fields.
This isn't everything, and I'm not even sure that it's safe or won't
conflict with something.  If it's a problem, I'll change it back and
I guess we'll live with the warnings at the C level.

* added a function rshare to the dict.h interface which allows you
to "copy" a dictionary in region r1 to a dictionary in region r2,
when r1 outlives r2.  The new dictionary shares the data in the
old dictionary but the new data are allocated in r2.  (I could use
this for the local bindings in tcenv, but decided to leave it with
the simpler linked list for now.)  To get the sharing to work out,
I have to declare the pointers in the tree to be "const".

* went through tcenv.cyc with a fine-toothed comb and made sure
that we were region-allocating everything possible, except for the
region partial order.