More changes to cut heap allocation or switch it to region-allocation.
* unboxed dictionaries.
* changed aggrdict_t to be an array of rval's -- no need to store the
  field name and avoids a lot of dictionary insertions at the price of
  some sharing.  Places have become a root and a list of ints, instead
  of a root and a list of fieldnames.
* in absyn, flattened Constraint, TunionInfoU, TunionFieldInfoU, and
  AggrInfoU. In tcpat, flattened Name_value.
* avoid allocating location info for pointers and tquals except when
  porting C code.
* toc translation environments now region-allocated.
* ptrsubtype now region-allocated
* tcArray now region-allocated
* more region allocation in Absynpp::to_tms
* check_catch_overlap in tcpat now region allocated
* construction of simple patterns in tcpat now region allocated
* environment and param-roots of cf analysis region allocated
* to_tmss in parser now region-allocates result

Before these changes when compiling parse_tab.cyc:
rgn_total_bytes: 3,909,280
heap_total_bytes: 13,420,190
heap_total_atomic_bytes: 496,564
number of collections: 9

After these changes:
rgn_total_bytes: 4,715,792
heap_total_bytes: 11,830,638
heap_total_atomic_bytes: 481,786
number of collections: 9


For reference, on the non-atomic heap allocation, parsing takes up
about 4.7MB, type-checking an additional 2.3MB, control-flow analysis
an additional 2.8MB, translation to C an additional 2MB, and dumpabsyn
about 100KB.

It's worth reminding you that type-checking does some AST allocation
(e.g., the types and topts that get put on the AST, resolved
identifiers and constants, etc.) but there's probably still a lot that
is temporary.  (Would be nice to know the live data size after a phase
to figure this out.)  Also, the parser is still allocating a lot of
temporary stuff including the tries in the lexer, and a lot of little
lists and so forth that don't neccessarily make it into the AST.  Toc
could be improved by sharing the identifiers that get allocated in
collapse_qvar_tag.  The dictionaries used in the control-flow analysis
are still the biggest single spot for heap- allocation.  In total,
dictionary operations chew up 2,128,924 bytes of the heap allocation,
and almost all of that is happening in the control-flow (other
possible culprits are the lexer dictionaries, rgn_order, and the
global dictionaries used in Toc, but I suspect only the former does a
lot of allocation.)

Added support in cycbison for threading some state from the parser
through the lexer (by defining some macros.)  As an example, the lbuf
is now created by the parser and passed in to each call of yylex so
that we don't use a global variable to store it.  This is a precursor
to moving all of the state of the lexer into an environment data
structure that can be effectively region-allocated.

Changed aprof so it doesn't print out all of the regions that were
empty at the beginning (I was getting swamped by these and we get
a report of those at the end anyway.)