Well, *I* didn't have to submit anything to PoPL :-)
Fixed leak in Dict::union_two_c where we were heap-allocating
something that should've been stack-allocated.

More changes to shift allocation from the heap into regions or
to get rid of allocation all-together.  Most of the shift is
due to region-allocating the stuff in the lexer.

Region-allocated the control-flow analysis.  This was a bitch, and I
still think that, in order to get it to scale properly, we need to
break each iteration of the analysis with a copying collection from
one region to another (doing a reset-region.)

Before region-allocating the cf analysis on parse_tab.cyc:

rgn_total_bytes: 7,034,192
heap_total_bytes: 9,367,936
heap_total_atomic_bytes: 485,394
number of collections: 6

After region-allocating the cf analysis:

rgn_total_bytes: 10,071,512
heap_total_bytes: 7,237,070
heap_total_atomic_bytes: 486,210
number of collections: 7

The running time for the compilation are almost identical (about 8.5
secs on my machine) though we're doing one more collection than
before.  Note that finally, we're doing more region allocation than
heap-allocation.  Also, I estimate that around 5 of the 7.7 MB
is abstract syntax that persists throughout compilation (so there's
no real reason to region-allocate it.)  So, I'm estimating that
we're region-allocating about 75% of what is possible.

Some things to note: this region-ification is pretty naive and
essentially puts all of the data for the cf analysis for a given
function in one region.  It also ends up copying some of that
information back out into the heap so that it can be attached to the
abstract syntax.  For instance, the annotations on expressions needed
to eliminate null & array-bounds checks get copied out each time
around the loop.  I would like to fix this so that we pass back a
separate data structure (e.g., hash-table) with this information.

Note that the statement flow annotations are *not* heap-allocated.
Instead, we put them in a hash-table and so they are collected after
the function is analyzed.  This saves us from having to do the
post-pass to clean up the annotations, at the price of looking things
up.