Changed syntax of alias stuff to be as Trevor suggested. I didn't do this before because the naive thing created lots of shift/reduce conflicts.  Here's the new syntax:
alias <`r>x = e in s

Adding the identifier 'in' in between e and s eliminates the
shift/reduce conflicts (it's parsed as an identifier and then
checked to be 'in' in the semantic actions).  To avoid the 'in'
I could have made s a compound statement, but this still had
two shift/reduce conflicts that I didn't feel like teasing out.