Added support for tracking non-variable locations for the uniqueness flow analysis.  It turns out that what was there before was unsound, since the following was considered well-typed:
  struct List<int,`U> l = List(1,unew List(1,NULL));
  struct List<int,`U> l2 = l;
  Core::ufree(l1.tl);
  Core::ufree(l2.tl);  // ouch!

We now consider aggregates (i.e. structs and tuples) that contain unique
pointers to be a so-called unique aggregates.  Whenever we are considering
something of struct or tuple type, we figure out all of the unique pointers
that can be directly reached from it.  In the above example for l, this
would be l.tl.  Then, whenever we assign a unique aggregate, all of its
directly reachable unique pointers are considered consumed.  In the above
example, the initialization of l2 causes l.tl to be consumed, and thus the
first Core::ufree() is disallowed.  This is implemented using the same
place_t stuff already present: places in the consumed set may now
consist of roots as well as roots with a list of fields indicating
the path to follow from the root to get to the place.

Like regular unique pointers, unique aggregates can only be used in consume
positions if they are reachable by a unique path, which at the moment means
not through any pointers, but only directly to the variable or constant, or
via a tuple- or struct-field access (recursively), as well as a couple of
other expression cases.

I added a bunch of test cases to tests/unique_test.cyc if you want to take a
look.

The next step is to allow tracking of locations accessed via pointers.  This
will be a little trickier but doable in at least some cases.