Added a basic form of typecase to Cyclone: you can now compare region handles for different regions and refine the region relationship as a result.  For example:
 int *`H varh;
 int *`U vary;

 void refine(region_t<`r::TR> r, int *`r::TR x) {
   if (r == Core::heap_region)
     varh = x;
   else if (r == Core::unique_region) {
     vary = x;
   let z = NULL;
   z :=: vary;
   Core::ufree(z);
 }

Region handles for different regions can be compared, but only if one is a
subkind of the other; this is to make sure that the resulting ordering
constraints make sense.

I only do the refinement for if statements at the moment, and only for
expressions that are equalities or inequalities; of course, there's no
reason we couldn't generalize to all test conditions (i.e. while loops,
conditional expressions, etc.) and all boolean expressions.

To implement this, in the typechecker I set region outlives constraints so
that each region outlives the other (in the body of the appropriate branch
of the if), and also change the kinds of the region variables to be the meet
of the two kinds.  I also had to change the runtime representation of the
unique_region: both the heap and unique regions were represented with a NULL
region handle, so the unique region is now (struct RegionHandle *)1.