I added support (modulo pattern matching) for the new form of tagged unions, i.e.,
@tagged union Foo { T1 x1; ... Tn xn; };

These are implemented similarly to the current flat datatypes (which I
expect to get rid of soon.)  Each tagged union includes a tag field.
A write to a member xi causes the tag to be updated with an integer i.
A read of member xi ensures that the tag has value i, throwing the
Match exception if not.

You build a tagged union by writing "Foo{.xi = e}" and you
pattern-match on a tagged union by writing "Foo{.xi = pat}".
However, I have not yet written the code needed for pattern-match
compilation.

The flow analysis can tell when you have not initialized at least one
member of a tagged union before using it, and it will give an error if
the member you are reading has not been initialized at some point.  It
allows a tagged union to escape (e.g., be passed to a function) as
long as one of the members has been written.

Note, however, that if you write x1 and then x2, the flow analysis
does not realize that x2 is no longer "initialized".  I achieved this
by adding a field (is_union) to the Aggregate r-value data structures
in the flow analysis, and using Aggregates for tagged unions (as well
as structs and tuples.)  Unlike structs and tuples, when we ask for
the initlevel of a union aggregate, it says the thing is AllIL (all
initialized) as long as one of the members is initialized.  I expect
Dan and Mike will want to look at this to see if this breaks any other
assumptions.  It would be nice to have the flow analysis make a member
uninitialized when some other member is written, but I was worried
this would cause the analysis to diverge.  (Dan?)

Restrictions:

Post-increment and decrement are not supported on members when the
resulting value is used (too much trouble to get the right value out
as a result.)  In other words, you can write "e.x++" in a statement
position, but not "(e.x++) + 4".

Due to Visual C++ constraints, you cannot initialize a top-level
tagged union with anything but the first member.  (VC++ does not have
support for designators, like GCC and C-9X have, and the only way
to initialize a union in VC is to use the first member.) I need
to fix the type-checker and code generator to deal with this issue.

Because of an unsoundness with aliases and updates to tagged unions
(see Dan's ESOP paper), you are not allowed to get a reference to one
of the tagged union members.  That is, you can't use "&e.xi" nor
"&e->xi", nor can you use a pattern of the form "Foo{.xi = *id}".

Still to do:

* pattern match compilation for tagged unions.

* top-level tagged union initializers can only use first member for VC++.

* add tagcheck(e.xi) which returns true xi was last member written.

* get flow analysis to better track the last member written

* use flow analysis info to avoid tag-checks/writes in translation.