Many changes which can be summarized as "replaced __flat__ datatypes with @tagged unions, and you can now use @fat, @thin, etc. as pointer qualifiers."
* Added pointer qualifiers:
@thin, @fat, @zeroterm, @nozeroterm, @region(t), @numelts(e), @notnull,
@nullable.  These can be mixed and matched with the original syntax,
but you'll have to do so carefully (see the pointer production in
the parser.)  Right now, if you put conflicting qualifiers in, then
you get no error message, and the last one (to the right) wins.
So, if you write "char * @zeroterm @nozeroterm" this is the
same as "char * @nozeroterm".

* Fixed a bug in the lexer with a hack.  When we ran into @taggedfoo,
this was lexed as "@tagged" "foo" instead of "@" "taggedfoo" which
caused occasional problems with Cyclone qualifiers.  So, I modified
the lexer to snarf "@tagged" and then if it sees an identifier
character following, it backs up the lexbuf.  (See
src/lex.cyl:cyc_qual.)  This seems to work, but I'm not happy with it.
Surely there's a better way?

* Fixed a bug in how tagged union declarations are handled within
namespaces.

* Fixed a bug in the resolution of unions and structs -- if you
declared "struct Foo {...}" and then wrote "union Foo" in a type
later, it would resolve to "struct Foo".  Instead, we now report
an error (as does GCC.)  I was a bit surprised to find out that
unions and structs don't have separate namespaces.

* Finished support for pattern matching on tagged unions.

* Added tagcheck(e.x) and tagcheck(e->x) where e is a tagged union.
This returns true when the last member written was x.  That is, it's
a way to test the tag without having to worry about an exception.
In the future, we can use tagcheck/pattern-matching info in the
flow analysis to avoid doing unnecessary tag checks.

* Aggregate patterns no longer need the name of the struct/tagged
union in front of the braces.  For instance, you can write:

let {.x = x, .y = y} = p;

instead of:

let Point{.x = x, .y = y} = p;

* Changed bison to use a @tagged union for the YYSTYPE instead of a
__flat__ datatype.  To keep things compatible, we now generate a
constructor for each member of the union, as well as a deconstructor.
Note, however, that if someone really wants to, they can do their own
type manipulations now.  That is, that could write $$, $1, etc., since
they can use the "." notation and their own union if they like.  So,
someday, we should perhaps provide a version of bison that works more
like the original.  One thing that changed is that you have to put in
all of the region annotations on the types (e.g., `H) to get things to
work out right with the constructors.

* Modified src/parse.y, tests/foo.cyc, tools/cyclex and lib/xml to get
them working with the changes to bison.

* Replaced all of the __flat__ datatypes with @tagged unions in the
rest of the compiler and then got rid of all of the code that dealt
with __flat__ datatypes all together, which simplified the
compiler somewhat.  In so doing, I triggered a very nasty bug which I
actually never could track down, though it had something to do with
the Constraint<`a> type in absyn.h.  Cutting down the test cases
shifted the bug and after spending a week on it, I've given up and
coded around it by centralizing more of the constraint manipulation.
I suspect that it was the infamous GCC bug, but am still not sure.
While debugging, I made a few other changes:

* Cut out a lot of the gentype stuff for now until we get it back
into shape.

* Changed the way we translate aggregate definitions (i.e., struct and
union declarations) to C.  Before, we were just mutating these in
place.  This meant that we were clobbering the original definition.  I
thought this might have some bad effect on the tagged union code, so I
changed it to not clobber the original definition and instead build
a new definition.  However, some of toc had grown to depend on the old
definition being mutated.  In particular, the checks to see if we need
to cast a field (e.g., because it's polymorphic) looked at the
translated type (instead of the original Cyclone type) to see if it
was void*.  I've changed this to see if it's void* or a type variable.
But be forewarned.