Lots of changes, in increasing order of importance:
* When the -stopafter-tc, -stopafter-cf, or -stopafter-parse flags are
specified, then we don't print out the cyc_include.h.

* Fixed bug in printing of (Cyclone) structs and unions -- was omitting
type parameters.

* You no longer have to put in explicit kinds for the type parameters of
datatype, struct, or union declarations when the kinds can be inferred
from the definition.  For instance, you can now write:

struct List<`a,`r> {`a hd; struct List<`a,`r> *`r tl;};

without having to put a ::R on the `r parameter.  As before, if you
have an abstract struct or datatype, then the kinds default to BoxKind
unless otherwise specified.

* A datatype is no longer associated with a region.  Instead,
you always use a pointer to a datatype to manipulate a datatype value,
and the region of the pointer determines where the datatype value
lives.  This means, in particular, that instead of writing:

  typedef datatype `r Foo foo_t<`r>

one will typically write:

  typedef datatype Foo @`r foo_t<`r>.

One nice side effect of this is that you can also write:

  typedef datatype Type * topt;

to represent a (possibly null) pointer to a type or even:

  typedef datatype Type ? type_array;

for an array of types and potentially avoid a level of indirection.
Note, however, that pattern matching is restricted to @ or * types.
You can pattern match on a datatype Foo ? value by coercing it to a
datatype Foo @/* before doing the match.

* As a type, datatype Foo now has kind A which means that it is treated
like an abstract struct -- something you can't directly dereference or
build, but rather something you have to manipulate through pointers.
A pointer to a datatype field (e.g., a value of type datatype Foo.C @)
is a subtype of datatype Foo @ and this is effectively the only way
to create a datatype value.  The only way to take apart a datatype
is through pattern matching.

* All datatype constructors are created (and matched) equally.
In particular, even if a data constructor has no arguments, then it
still must be allocated somehwere to be manipulated as a value of the
datatype, and when matching a datatype value, each constructor will
have the "&" pattern.  On the plus side, this makes things a bit more
consistent and greatly simplified hunks of the compiler where we had
special-case code for nullary constructors.  On the minus side, it
means that pattern matching isn't quite as efficient [although it's
not so bad since we no longer have to do tests to see separate out
nullary constructors from value-carrying ones] and you are responsible
for allocating nullary constructors (see below).

* One side effect of the change is that if you want to use equality
tests on nullary constructors (e.g., in the compiler we might have
written t == VoidType) then you'll have to declare a global variable
to hold the VoidType value, and use its address consistently.  For
instance:

  datatype Type.VoidType VoidType_val = VoidType;
  datatype Type @void_typ = &VoidType_val;

in which case the test above would be replaced with t == void_typ.
Within the compiler, I used two macros (one for interfaces and one
for implementations) to simplify the declaration of these shared
constructors.

* Changed as many datatypes to enums as possible (i.e., when all the
constructors were nullary).  That allowed me to avoid declaring so
many of the globals described above.

* Made enum's boxed types.  Not necessarily a good idea for
portability's sake, but for the moment, it made it easier to convert
more of our datatypes to enums.

* Changed the pattern match compiler so that it does less
casting, and so that it's a little easier to see which
case you're in.  In particular, the code that gets generated
for a datatype constructor match case Foo(x1,x2,x3) looks
like this now:

  struct Cyc_Foo *temp = (struct Cyc_Foo *)v;
  if (temp->tag != 3) goto L;
  x1 = temp->f1;
  x2 = temp->f2;
  x3 = temp->f3;

where it looked more like this before:

  if (*((int *)v) != 3) goto L;
  x1 = ((struct Cyc_Foo *)v)->f1;
  x2 = ((struct Cyc_Foo *)v)->f2;
  x3 = ((struct Cyc_Foo *)v)->f3;

The OKE folks requested that we make this a bit more readable
to help with debugging the output of the compiler.