Various changes to get mini-httpd to compile.  In particular, I added back the struct sockaddr that was previously omitted, along with struct stat.  Cygwin does some funny things with struct stat, so I had to add in a #ifdef to get things to work.
Also, the grp library was using a function to convert a zero-
terminated array of C-strings to a mstring_t?.  That function
used to live in runtime_cyc.c but is now commented out, since
we can do this from within the language.  So, I fixed the library
to do this.  I also ripped this function ntCsl_to_ntsl out
of core.h so no one will be tempted to use it again.