"make nocheck" seems to be broken -- this is needed for tests/shootout.
Modified the treatment of unique pointers in the flow analysis.  In
essence, I folded the "consumed" information into the flowdict.  This
simplified things and allowed me to be quite a bit more precise about
where unique pointers are consumed.  Here's how it works in a bit more
detail:

First, the analysis now tracks whether or not it is in a "copy
context".  It only worries about consuming unique pointer values
within a copy context.  That simplified a number of issues.

Second, when you make a copy of a unique pointer, we conceptually make
the "old" path to the pointer uninitialized.  This means that there
has to be an old path.  To accomplish this, we unfold unique paths as
needed by adding MallocPt nodes corresponding to the dererence points
into the flowdict.  For instance, if you have a unique list x and do
x->tl->tl, then we'll add MallocPt nodes for the ->tl and ->tl->tl
expressions.  The ->tl->tl will then become uninitialized (assuming
it's being copied.)

The nice thing about this is that you don't have to do swaps to
get to unique pointers, as long as they are grabbed along unique
paths.  Rather, you can just read the pointer out and the analysis
expands the "frontier" necssary to know that the old copy has
been consumed.

Third, in practice, we can't just make these things uninitialized
(Unknown(NoneIL)) because of the local iteration done for the unknown
order of evaluation.  Instead, we add a new abstract Rvalue called
Consume(e,i,oldrval) which represents a node that has been consumed by
expression e in iteration i of the analysis.  When doing a local
fixed-point computation (e.g., for a function call), the same
expression is allowed to read out the old rval as if it was not
consumed, but all other expressions have to treat this consumed value
as if it was Unknown(NoneIL).

Fourth, we've added a new attribute noliveunique(i) which is a lousy
name, but I couldn't think of a better one.  It's used for ufree to
specify that the parameter is unique, but what it points to shouldn't
have any (usable) unique pointers in it.  For instance, if you have:

  struct P {int @`U x; int@`U y};

  void foo(struct P@`U p) {
    ufree(p);
  }

then the call to ufree(p) will give you a warning that you might be
losing references to some unique pointers (in this case p->x and
p->y.)  If instead, you did:

  void foo(struct P@`U p) {
    ufree(p->x);
    ufree(p->y);
    ufree(p);
  }

then you don't get any warnings because none of the objects passed to
ufree contain any usable unique pointers in them (and the analysis
knows this.)

I decided to make this an attribute instead of baking it into to ufree
so that you could lift it to other functions.

There are still some kinks to work out.  First, the alias construct is
broken due to my changes.  That meant that I had to change the
shootout benchmarks which used "alias" to use regions instead of
unique pointers for now.  Second, joins on expanded nodes in the
analysis don't work as well as they could, since the addresses
generated don't match up.  We'd be better off having "anonymous" nodes
when we expand unique pointers (i.e., we'd get more precision in the
analysis.)  Third, I have not yet checked that I'm treating
polymorphic types properly.  Mike's recent changes to get uniqueness
and polymorphism squared away make this possible.  Fourth, some
warnings are misleading when you ufree a possibly non-null unique
pointer.  And finally, we still don't warn when you lose a unique
pointer due to a join, or exiting the scope of a variable whose value
might contain unique pointers.