The dependencies are not working properly for "make cyclone_src". In particular, if I touch cf_flowinfo.h, then it recompiles the .c files for the cyclone files that depend upon it, but not the .cyc files as it should.  Haven't investigated this...
Fixed the control flow analysis w.r.t. the alias construct, though
there's a bit of a hack in there that essentially assumes that
the expression being aliased is of the form "(T)e" (i.e., a cast.)
I hope to get rid of this by cleaning up the alias syntax and
forcing the user to put the type in explicitly.

Also, I got rid of a spurious warning that was generated when you
tried to ufree a possibly null pointer.