New flag -Wlose-unique which tries to warn in some cases where you might lose a unique pointer.  In particular, it uses the flow analysis to determine whether or not a local variable or parameter still might contain a valid unique pointer when you exit its scope.  This is not entirely easy due to the non-local control flow.  This isn't as precise as I'd hoped because when we do a non-local control transfer, we have to crawl through the flowdicts at the join point to figure out what has gone out of scope.  I was worried that this might be expensive, so I have turned off this flag by default.
I still need to add code to warn when you overwrite a unique pointer.
Note also that you can always lose a unique pointer by putting it into
a shared data structure and then lose all references to the shared
data structure.  (That's the price we pay for affine pointers.)