Minor fix:
For the most part, the flow-analysis determines if a function may "fall
off" without returning.  That's fine for return-type void, a warning
for integral return types, and an error otherwise.

But for obscure reasons, it's the type-checker that notices when a
break statement causes a function to immediately "fall off" (e.g., the
break is in a loop that is the function's last statement).  And that
code was making such a break an error for integral return types.

Now it's a warning, just like in the flow analysis.

(It took longer to describe than it took to fix.)